+++ src/main/java/net/brnbrd/delightful/compat/jade/DCropProgress.java
@@ -17,9 +17,10 @@ public enum DCropProgress implements IBlockComponentProvider {
 	INSTANCE;
 
 	private static void addMaturityTooltip(ITooltip tooltip, int age, int maxAge) {
-		tooltip.add(Component.translatable("tooltip.jade.crop_growth", age == maxAge ?
-				Component.translatable("tooltip.jade.crop_mature").withStyle(ChatFormatting.GREEN) :
-				Component.literal(String.format("%.0f%%", (age / (float) maxAge) * 100.0F)).withStyle(ChatFormatting.WHITE)
+		tooltip.add(Component.translatable("tooltip.jade.crop_growth",
+			age == maxAge ?
+			Component.translatable("tooltip.jade.crop_mature").withStyle(ChatFormatting.GREEN) :
+			Component.literal(String.format("%F%%", (age / (float) maxAge) * 100F)).withStyle(ChatFormatting.WHITE)
 		));
 	}